+++ front-end/bdas2-frontend/src/views/entities/Spoje/SpojDetail.vue
@@ -2,13 +2,18 @@
   <div class="main-wrapper">
     <MainHeader title="Detail spoje"/>
     <form>
+      <div v-if="error" class="alert alert-danger">{{error}}</div>
       <div class="mb-3">
         <label for="nazev" class="form-label">Název</label>
-        <input v-model="spojModel.NAZEV_SPOJ" type="text" class="form-control" id="nazev">
+        <input v-model="spojModel.NAZEV_SPOJ" type="text" class="form-control" id="nazev" :class="{ 'is-invalid': submitted && !spojModel.NAZEV_SPOJ }">
+        <div v-show="submitted && !spojModel.NAZEV_SPOJ" class="invalid-feedback">Název je povinný</div>
       </div>
-      <select-box @input="updateSelectedJROption" label="Vyberte jízdní řád" :options="radyOptions" :selected="spojModel.ID_JIZDNI_RAD"/>
-      <select-box @input="updateSelectedOption" label="Vyberte linku" :options="linkyOptions" :selected="spojModel.ID_LINKA"/>
-      <double-time-form-group v-for="(zastavka, index) in zastavky" :key="index" :item="zastavka"/>
+      <select-box @input="updateSelectedJROption" label="Vyberte jízdní řád" :options="radyOptions" :selected="spojModel.ID_JIZDNI_RAD"  :class="{ 'is-invalid': submitted && !spojModel.ID_JIZDNI_RAD }" />
+      <div v-show="submitted && !spojModel.ID_JIZDNI_RAD" class="invalid-feedback">Zvolte jízdní řád</div>
+      <select-box @input="updateSelectedOption" label="Vyberte linku" :options="linkyOptions" :selected="spojModel.ID_LINKA" :class="{ 'is-invalid': submitted && !spojModel.ID_LINKA }" />
+      <div v-show="submitted && !spojModel.ID_LINKA" class="invalid-feedback">Zvolte linku, kterou bude spoj popisovat</div>
+      <double-time-form-group v-for="(zastavka, index) in zastavky" :key="index" :item="zastavka" :class="{ 'is-invalid': submitted && !doesEveryStopHaveTimes }" />
+      <div v-show="submitted && !doesEveryStopHaveTimes" class="invalid-feedback">Nastavte časy u všech zastávek</div>
       <button :disabled="areDataSame" @click="handleUpravit" class="btn btn-primary">Upravit</button>
     </form>
   </div>
@@ -40,10 +45,10 @@ export default {
         ID_JIZDNI_RAD: null
       },
       spoj: {},
-      jizdniRad: null,
-      prevJizdniRad: null,
       zastavky: [],
       prevZastavky: [],
+      submitted: false,
+      error: ''
     }
   },
   created () {
@@ -101,45 +106,60 @@ export default {
     },
     handleUpravit () {
       event.preventDefault();
-      if (!this.areObjectsEqual(this.spoj, this.spojModel)) {
-        SpojService.updateEntity(this.getIdDetail, {
-          NAZEV_SPOJ: this.spojModel.NAZEV_SPOJ,
-          ID_LINKA: this.spojModel.ID_LINKA
-        });
-      }
-      if (this.spoj.ID_LINKA === this.spojModel.ID_LINKA) {
-        this.prevZastavky.forEach((zaznam) => {
-          ZaznamjizdnihoraduService.deleteEntity(zaznam.ID_ZAZNAM_JIZDNIHO_RADU).catch((e) => {console.log("něco se pokazilo")})
-        });
-        this.zastavky.forEach((zaznam) => {
-          ZaznamjizdnihoraduService.createEntity({
-            ID_SPOJ: this.getIdDetail,
-            ID_JIZDNI_RAD: this.spojModel.ID_JIZDNI_RAD,
-            ID_ZASTAVKA_LINKA: zaznam.ID_ZASTAVKA_LINKA,
-            PRAVIDELNY_PRIJEZD: zaznam.PRAVIDELNY_PRIJEZD,
-            PRAVIDELNY_ODJEZD: zaznam.PRAVIDELNY_ODJEZD
+      this.submitted = true;
+      if (this.isOK) {
+        if (!this.areObjectsEqual(this.spoj, this.spojModel)) {
+          SpojService.updateEntity(this.getIdDetail, {
+            NAZEV_SPOJ: this.spojModel.NAZEV_SPOJ,
+            ID_LINKA: this.spojModel.ID_LINKA
           });
-        })
-      } else if (this.spoj.ID_JIZDNI_RAD !== this.spojModel.ID_JIZDNI_RAD
-          || !this.areObjectsArrayEqual(this.zastavky, this.prevZastavky)) {
-        this.zastavky.forEach((zaznam) => {
-          ZaznamjizdnihoraduService.updateEntity(zaznam.ID_ZAZNAM_JIZDNIHO_RADU, {
-            ID_SPOJ: this.getIdDetail,
-            ID_JIZDNI_RAD: this.spojModel.ID_JIZDNI_RAD,
-            ID_ZASTAVKA_LINKA: zaznam.ID_ZASTAVKA_LINKA,
-            PRAVIDELNY_PRIJEZD: zaznam.PRAVIDELNY_PRIJEZD,
-            PRAVIDELNY_ODJEZD: zaznam.PRAVIDELNY_ODJEZD
-          }, 'ID_ZAZNAM_JIZDNIHO_RADU')
-        });
-        this.init();
+        }
+        if (this.spoj.ID_LINKA === this.spojModel.ID_LINKA) {
+          this.prevZastavky.forEach((zaznam) => {
+            ZaznamjizdnihoraduService.deleteEntity(zaznam.ID_ZAZNAM_JIZDNIHO_RADU).catch((e) => {console.log("něco se pokazilo")})
+          });
+          this.zastavky.forEach((zaznam) => {
+            ZaznamjizdnihoraduService.createEntity({
+              ID_SPOJ: this.getIdDetail,
+              ID_JIZDNI_RAD: this.spojModel.ID_JIZDNI_RAD,
+              ID_ZASTAVKA_LINKA: zaznam.ID_ZASTAVKA_LINKA,
+              PRAVIDELNY_PRIJEZD: zaznam.PRAVIDELNY_PRIJEZD,
+              PRAVIDELNY_ODJEZD: zaznam.PRAVIDELNY_ODJEZD
+            });
+          })
+        } else if (this.spoj.ID_JIZDNI_RAD !== this.spojModel.ID_JIZDNI_RAD
+            || !this.areObjectsArrayEqual(this.zastavky, this.prevZastavky)) {
+          this.zastavky.forEach((zaznam) => {
+            ZaznamjizdnihoraduService.updateEntity(zaznam.ID_ZAZNAM_JIZDNIHO_RADU, {
+              ID_SPOJ: this.getIdDetail,
+              ID_JIZDNI_RAD: this.spojModel.ID_JIZDNI_RAD,
+              ID_ZASTAVKA_LINKA: zaznam.ID_ZASTAVKA_LINKA,
+              PRAVIDELNY_PRIJEZD: zaznam.PRAVIDELNY_PRIJEZD,
+              PRAVIDELNY_ODJEZD: zaznam.PRAVIDELNY_ODJEZD
+            }, 'ID_ZAZNAM_JIZDNIHO_RADU')
+          });
+        }
+      setTimeout(() => { this.init(); }, 1000);
+      } else {
+        this.error = 'Doplňte potřebné údaje';
       }
     }
   },
   computed: {
     areDataSame () {
-      console.log("are data same?", this.prevZastavky, this.zastavky);
       return this.areObjectsEqual(this.spoj, this.spojModel) && this.areProxyObjectsArrayEqual(this.prevZastavky, this.zastavky);
-    }
+    },
+    isOK () {
+      return !!this.spojModel.NAZEV_SPOJ && this.spojModel.ID_LINKA && this.spojModel.ID_JIZDNI_RAD && this.doesEveryStopHaveTimes;
+    },
+    doesEveryStopHaveTimes () {
+      for (const key in this.zastavky) {
+        if (!this.zastavky[key].PRAVIDELNY_PRIJEZD || !this.zastavky[key].PRAVIDELNY_ODJEZD) {
+          return false;
+        }
+      }
+      return true;
+    },
   }
 }
 </script>
